+++ plugins/main/public/components/overview/common/initial-width.ts
@@ -1,0 +1,7 @@
+export const commonInitialWidth = {
+  timestamp: 203.8,
+  'agent.id': 93.05,
+  'agent.name': 318.05,
+  'rule.level': 93.05,
+  'rule.id': 93.05,
+} as const;