+++ plugins/main/public/components/overview/threat-hunting/events/threat-hunting-columns.tsx
@@ -1,62 +1,41 @@
 import { EuiDataGridColumn } from '@elastic/eui';
 import { tDataGridColumn } from '../../../common/data-grid';
+import { commonColumns } from '../../common/data-grid-columns';
 
 export const MAX_ENTRIES_PER_QUERY = 10000;
 
-export const threatHuntingTableDefaultColumns: tDataGridColumn[] = [
-  {
+const threatHuntingTableCommonColumns = {
+  icon: {
     id: 'icon',
   },
-  {
-    id: 'timestamp',
-    isSortable: true,
-    defaultSortDirection: 'desc',
-  },
-  {
-    id: 'agent.id',
-  },
-  {
-    id: 'agent.name',
-  },
-  {
+  'rule.mitre.id': {
     id: 'rule.mitre.id',
+    initialWidth: 123,
   },
-  {
+  'rule.mitre.tactic': {
     id: 'rule.mitre.tactic',
+    initialWidth: 266,
   },
-  {
-    id: 'rule.description',
-  },
-  {
-    id: 'rule.level',
-  },
-  {
-    id: 'rule.id',
-  },
+} as Record<string, tDataGridColumn>;
+
+export const threatHuntingTableDefaultColumns: tDataGridColumn[] = [
+  threatHuntingTableCommonColumns.icon,
+  commonColumns.timestamp,
+  commonColumns['agent.id'],
+  commonColumns['agent.name'],
+  threatHuntingTableCommonColumns['rule.mitre.id'],
+  threatHuntingTableCommonColumns['rule.mitre.tactic'],
+  commonColumns['rule.description'],
+  commonColumns['rule.level'],
+  commonColumns['rule.id'],
 ];
 
 export const threatHuntingTableAgentColumns: EuiDataGridColumn[] = [
-  {
-    id: 'icon',
-  },
-  {
-    id: 'timestamp',
-    isSortable: true,
-    defaultSortDirection: 'desc',
-  },
-  {
-    id: 'rule.mitre.id',
-  },
-  {
-    id: 'rule.mitre.tactic',
-  },
-  {
-    id: 'rule.description',
-  },
-  {
-    id: 'rule.level',
-  },
-  {
-    id: 'rule.id',
-  },
+  threatHuntingTableCommonColumns.icon,
+  commonColumns.timestamp,
+  threatHuntingTableCommonColumns['rule.mitre.id'],
+  threatHuntingTableCommonColumns['rule.mitre.tactic'],
+  commonColumns['rule.description'],
+  commonColumns['rule.level'],
+  commonColumns['rule.id'],
 ];